requirements.txt
anthropic
Pillow
python-dotenv
